Add wsgi and create terms and condition
Add wsgi file, create page for about terms privacy usage and all that.
Also modify for header to not show usual header if the person is not log
in. Comment will ask user to log in if not log in. Also add script to
just press enter to search user. Fix user search bug.